requirements.txt
django==5.2.7
pillow==11.3.0
django-imagekit==6.0.0
pillow-heif==1.1.1
python-dotenv>=1.0,<2.0
